feat: add cluster and upload btns for config panels
