+++ frontend/src/components/AdvancedDrawer.vue
@@ -14,7 +14,12 @@
           <KyvernoConfig />
         </v-list-item>
         <v-list-item>
-          <AdvancedConfigDialog id="crd" title="Custom Resource Definitions" :info="options.panels.crdInfo" v-model="inputs.customResourceDefinitions" />
+          <AdvancedConfigDialog id="crd" title="Custom Resource Definitions" :info="options.panels.crdInfo" v-model="inputs.customResourceDefinitions">
+            <template #actions="{ update, content }">
+              <UploadButton label="Upload" @click="update" :tooltip="false" />
+              <ClusterCRDButton @update:model-value="update" :model-value="content" label="From Cluster" />
+            </template>
+          </AdvancedConfigDialog>
         </v-list-item>
         <v-list-item>
           <AdvancedConfigDialog
@@ -22,12 +27,18 @@
             title="Policy Exceptions"
             :info="options.panels.exceptionsInfo"
             v-model="inputs.exceptions"
-            uri="policyexception.yaml" />
+            uri="policyexception.yaml">
+            <template #actions="{ update, content }">
+              <UploadButton label="Upload" @click="update" :tooltip="false" />
+              <ClusterExceptionButton @update:model-value="update" :model-value="content" />
+            </template>
+          </AdvancedConfigDialog>
         </v-list-item>
         <v-list-item>
           <AdvancedConfigDialog id="clusterResources" title="Cluster Resources" :info="options.panels.clusterResourcesInfo" v-model="inputs.clusterResources">
-            <template #append="{ update, content }">
-              <ClusterResourceButton @update:model-value="update" :model-value="content" />
+            <template #actions="{ update, content }">
+              <UploadButton label="Upload" @click="update" :tooltip="false" />
+              <ClusterResourceButton @update:model-value="update" :model-value="content" label="From Cluster" />
             </template>
           </AdvancedConfigDialog>
         </v-list-item>
@@ -48,6 +59,9 @@ import AdvancedConfigDialog from './Advanced/AdvancedConfigDialog.vue'
 import KyvernoConfig from './Advanced/KyvernoConfig.vue'
 import { inputs } from '@/store'
 import ClusterResourceButton from './Panel/ClusterResourceButton.vue'
+import ClusterExceptionButton from './Panel/ClusterExceptionButton.vue'
+import UploadButton from './Panel/UploadButton.vue'
+import ClusterCRDButton from './Panel/ClusterCRDButton.vue'
 
 const props = defineProps({
   modelValue: { type: Boolean, default: false }
